refactor: Update NavBar component to add language selection buttons
This commit updates the SVG assets for the language buttons in the NavBar component. The "gb.svg" and "el.svg" files have been added to the "src/assets" directory. These SVGs are used as background images for the language buttons in the NavBar component.

The NavBar.js file has been modified to include the language buttons and their corresponding functionality. The "LngBtnContainer" div has been added to hold the buttons, and the "LngBtn" class has been defined in the NavBar.module.css file to style the buttons. The buttons are now displayed with the Greek and English flags as background images, and clicking on them triggers the change of the language using the i18n module.

These changes improve the user experience by providing language options and making the language selection more visually appealing.